Moves Map-related enums to Sandbox.
