+++ ParquetClassLibrary/Sandbox/ChunkOrientation.cs
@@ -1,4 +1,4 @@
-namespace ParquetClassLibrary.Sandbox.IDs
+namespace ParquetClassLibrary.Sandbox
 {
     /// <summary>
     /// Indicates a the major flow of parquets procedurally generated in a MapChunk.
